+++ src/Avalonia.Controls/Notifications/Notification.cs
@@ -1,15 +1,29 @@
-﻿using System;
+﻿// Copyright (c) The Avalonia Project. All rights reserved.
+// Licensed under the MIT license. See licence.md file in the project root for full license information.
+
+using System;
 
 namespace Avalonia.Controls.Notifications
 {
     /// <summary>
-    /// Defines content for a <see cref="NotificationCard"/> control.
+    /// Implements the INotification interfaces.
+    /// Can be displayed by both <see cref="INotificationManager"/> and <see cref="IManagedNotificationManager"/>
     /// </summary>
     /// <remarks>
     /// This notification content type is compatible with native notifications.
     /// </remarks>
     public class Notification : INotification
     {
+        /// <summary>
+        /// Instantiates an instance of <see cref="Notification"/>
+        /// </summary>
+        /// <param name="title">The title of the notification.</param>
+        /// <param name="message">The message to be displayed in the notification.</param>
+        /// <param name="type">The <see cref="NotificationType"/> of the notification.</param>
+        /// <param name="expiration">The expiry time at which the notification will close. 
+        /// Use <see cref="TimeSpan.Zero"/> for notifications that will remain open.</param>
+        /// <param name="onClick">The Action to call when the notification is clicked.</param>
+        /// <param name="onClose">The Action to call when the notification is closed.</param>
         public Notification(string title,
             string message,
             NotificationType type = NotificationType.Information,
@@ -25,16 +39,22 @@ namespace Avalonia.Controls.Notifications
             OnClose = onClose;
         }
 
+        /// <inheritdoc/>
         public string Title { get; private set; }
 
+        /// <inheritdoc/>
         public string Message { get; private set; }
 
+        /// <inheritdoc/>
         public NotificationType Type { get; private set; }
 
+        /// <inheritdoc/>
         public TimeSpan Expiration { get; private set; }
 
+        /// <inheritdoc/>
         public Action OnClick { get; private set; }
 
+        /// <inheritdoc/>
         public Action OnClose { get; private set; }
     }
 }
